Refactor socket listeners
- move socket event handlers to separate module  (use dependency
injection to make unit testing easier)
- add basic CSS reset and some basic button styles
