+++ client/index.html
@@ -12,7 +12,7 @@
 
     <form>
       <input id="user-message" name="message"/>
-      <button type="submit">Send</button>
+      <button class="btn btn--primary"type="submit">Send</button>
     </form>
 
     <ul id="active-users"></ul>
