+++ client/scss/components/_buttons.scss
@@ -1,7 +1,31 @@
 .btn {
   display: inline-block;
+  padding: 13px 25px;
+  border: 2px solid getColor(secondary);
+  border-radius: 30px;
+  will-change: background-color, color;
+  text-transform: uppercase;
+  font-size: 14px;
+  font-weight: bold;
+  white-space: nowrap;
+  transition: background-color 0.2s ease-in-out, color 0.2s ease-in-out;
+
+  :disabled {
+    cursor: not-allowed;
+    opacity: 0.5;
+  }
 
   &--primary {
-    color: blue;
+    background-color: transparent;
+    color: getColor(primary);
+
+    &:hover {
+      background-color: getColor(secondary);
+      color: getColor(white);
+    }
+  }
+
+  &--secondary {
+    color: getColor(white);
   }
 }
