+++ client/scss/utils/_variables.scss
@@ -13,6 +13,8 @@ $colors: (
   accent: #b09a68,
   dark: #080709,
   dark-lighten: #999,
+  primary: #615ffa,
+  secondary: #7510f7,
 );
 
 $layers: (
