+++ server/app.js
@@ -3,6 +3,7 @@ import socketIO from 'socket.io';
 import path from 'path';
 import EVENTS from '../common/socket-events';
 import usersRepo from './users-repository';
+import socketControllerFactory from './socket-controller';
 
 export const initApp = (app) => {
   const http = createServer(app);
@@ -13,67 +14,21 @@ export const initApp = (app) => {
     res.sendFile(path.resolve(__dirname, '../client/index.html'));
   });
 
-  io.on('connection', (socket) => {
-    // State of the current user
-    const userState = {
-      isAdded: false,
-    };
+  io.on('connect', (socket) => {
+    console.log('a new user (socket) connected');
+    const socketController = socketControllerFactory(socket, usersRepository);
 
-    console.log('a user connected');
-    console.log(usersRepository.getUsers());
-    socket.on('disconnect', () => {
-      if (userState.isAdded) {
-        usersRepository.deleteUser(socket.id);
+    // Setup socket disconnection listener
+    socket.on('disconnect', socketController.handleDisconnect);
 
-        // Emit to other users that current user has disconnected
-        socket.broadcast.emit(EVENTS.USER_DISCONNECTED, {
-          socketId: socket.id,
-          userName,
-        });
-      }
-      console.log('user disconnected');
-    });
-
-    socket.on(EVENTS.USER_ADD, (userName) => {
-      userName = userName.trim();
-
-      if (userState.isAdded) {
-        return socket.emit(EVENTS.USER_EXISTS, { userName: socket.userName });
-      } else if (usersRepository.isUserNameAlreadyTaken(userName)) {
-        return socket.emit(EVENTS.USER_NAME_ALREADY_TAKEN, { userName });
-      }
-
-      socket.userName = userName;
-
-      // Add new user to the list of active users
-      const createdUser = usersRepository.createUser(
-        socket.id,
-        socket.userName
-      );
-
-      userState.isAdded = true;
-
-      // Emit global event with app state details
-      socket.broadcast.emit(EVENTS.USER_JOINED, {
-        numberOfUsers: usersRepository.getNumberOfUsers(),
-        createdUser: createdUser,
-      });
-
-      // Emit an event (only to current socket) after current user is logged in
-      socket.emit(EVENTS.USERS_INITIAL_UPDATE, {
-        userName: socket.userName,
-        numberOfUsers: usersRepository.getNumberOfUsers(),
-        activeUsers: usersRepository.getUsersWithoutCurrentClient(socket.id),
-      });
-
-      console.log('users', usersRepository.getState());
-    });
+    // Setup adding new user listener
+    socket.on(EVENTS.USER_ADD, socketController.handleUserAdd);
   });
 
   const PORT = process.env.PORT || 3000;
 
   http.listen(PORT, () => {
-    console.log(`Example app listening at http://localhost:${PORT}`);
+    console.log(`Socket chat app listening at http://localhost:${PORT}`);
   });
 
   return app;
